requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
pydantic
python-multipart
python-dotenv
sqlalchemy
pypdf
python-docx
sentence-transformers
requests
